:racehorse: :construction: Improve freq. selection in Nyquist analysis
If the user does not provide a frequency vector, then the Nyquist
analysis now selects the frequencies based on an adaptive
algorithm. This algorithm adjust the step to provide a smooth plot.

Initial tests have been performed, but a more extensive set of
tests is needed to verify if everything is working as expected.